Change parameter style of the validator
